Moved "sortArrayOfObjectsByKey" to util.js
