+++ src/components/film-list-extra.js
@@ -1,32 +1,11 @@
 import {createFilmCard} from "../components/film-card";
+import {sortArrayOfObjectsByKey} from "../util/util";
 
 let SECTION_NAMES = new Map([
   [`rating`, `Top Rated`],
   [`comments`, `Most commented`]
 ]
 );
-
-const sortArrayOfObjectsByKey = (key, order = `desc`) => {
-  return function innerSort(a, b) {
-    if (!a.hasOwnProperty(key) || !b.hasOwnProperty(key)) {
-      // property doesn't exist on either object
-      return 0;
-    }
-
-    const varA = (typeof a[key] === `string`) ? a[key].toUpperCase() : a[key];
-    const varB = (typeof b[key] === `string`) ? b[key].toUpperCase() : b[key];
-
-    let comparison = 0;
-    if (varA > varB) {
-      comparison = 1;
-    } else if (varA < varB) {
-      comparison = -1;
-    }
-    return (
-      (order === `desc`) ? (comparison * -1) : comparison
-    );
-  };
-};
 
 const createFilmCards = (films, quantity) => {
   let markup = ``;
